// src/@v2/forms/application/system-ai-prompt/constants/system-ai-prompt-key.enum.ts
export enum SystemAiPromptKeyEnum {
  RISK_SOURCES_RECOMMENDATIONS = 'RISK_SOURCES_RECOMMENDATIONS',
  RISK_NARRATIVE_DIAGNOSTIC = 'RISK_NARRATIVE_DIAGNOSTIC',
  INDICATORS_NARRATIVE_ANALYSIS = 'INDICATORS_NARRATIVE_ANALYSIS',
  INDICATORS_NARRATIVE_DIAGNOSTIC = 'INDICATORS_NARRATIVE_DIAGNOSTIC',
  INDICATORS_NARRATIVE_DIAGNOSTIC_GROUPS_ONLY = 'INDICATORS_NARRATIVE_DIAGNOSTIC_GROUPS_ONLY',
  INDICATORS_NARRATIVE_DIAGNOSTIC_GROUPS_AND_QUESTIONS = 'INDICATORS_NARRATIVE_DIAGNOSTIC_GROUPS_AND_QUESTIONS',
  RISK_FACTOR_CHEMICAL_AI_SUGGESTIONS = 'RISK_FACTOR_CHEMICAL_AI_SUGGESTIONS',
  RISK_FACTOR_PHYSICAL_AI_SUGGESTIONS = 'RISK_FACTOR_PHYSICAL_AI_SUGGESTIONS',
  RISK_FACTOR_BIOLOGICAL_AI_SUGGESTIONS = 'RISK_FACTOR_BIOLOGICAL_AI_SUGGESTIONS',
  RISK_FACTOR_ERGONOMIC_AI_SUGGESTIONS = 'RISK_FACTOR_ERGONOMIC_AI_SUGGESTIONS',
  RISK_FACTOR_ACCIDENT_AI_SUGGESTIONS = 'RISK_FACTOR_ACCIDENT_AI_SUGGESTIONS',
  RISK_FACTOR_OTHER_AI_SUGGESTIONS = 'RISK_FACTOR_OTHER_AI_SUGGESTIONS',
}